Fixed #3 -- now works, sans output formatting

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -5,13 +5,13 @@ import com.techelevator.tenmo.model.Account;
 import java.math.BigDecimal;
 
 public interface AccountDao {
-    BigDecimal getBalance(int userId);
+    BigDecimal getBalance(long userId);
 
-    BigDecimal addToBalance(BigDecimal amountToAdd, int id);
+    BigDecimal addToBalance(BigDecimal amountToAdd, long id);
 
-    BigDecimal subtractFromBalance(BigDecimal amountToSubtract, int id);
+    BigDecimal subtractFromBalance(BigDecimal amountToSubtract, long id);
 
-    Account findUserById(int userId);
+    Account findUserById(long userId);
 
-    Account findAccountById(int id);
+    Account findAccountById(long id);
 }
